+++ CHANGELOG.md
@@ -13,6 +13,8 @@
   broken over multiple lines.
 - The formatter style has been improved to wrap lists and bit strings over as
   few lines as possible if the elements are Ints, Floats, or Strings.
+- The formatter style has been improved to preserve comments on labelled
+  call arguments.
 - The performance of the formatter has been improved.
 - Records can be updated using the spread syntax. A warning is emitted if no
   fields are updated when using this syntax.